Use UTF-8 for stdout, except TTY
UTF-8 is desirable in most situations, particularly if stdout is
directed to a file (possibly after piped post-processing) or piped to an
XML processor.  It is not desirable when stdout is a TTY which
interprets the output without regard for the XML declaration.  For this
case, use the encoding the terminal expects, as determined by the Python
runtime.

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>