pylintrc: disable unspecified-encoding
When open() is called without an encoding argument, unspecified-encoding
produces the following warning:

    W1514: Using open without explicitly specifying an encoding (unspecified-encoding)

This is difficult to work around for projects which support Python 2,
where open() does not accept an encoding argument.  Although it would be
possible to use `io.TextIOWrapper(io.BufferedReader(io.FileIO()))`, it
would produce `unicode` instead of `bytes`, which is often undesirable.
Ignore the error in projects supporting Python 2.

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>