drop pylint_runner
It appears that pylint lints packages recursively, so it is only
necessary to specify each top-level package/directory and standalone
files.  For many projects, this is sufficient.  Drop the pylint_runner
dependency and call pylint directly.

Note: This also avoids a potential blocker for moving from .pylintrc to
pyproject.toml, which is not currently supported by pylint_runner for
reading the ignores list.

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>