test: mock os.fdopen for argcomplete
argcomplete opens file descriptor 9 for writing debugging and error
messages.  If file descriptor 9 is already open, as it is when running
under pytest, it is reopened and subsequently close is attempted twice,
causing an error.

To avoid this, mock os.fdopen.  Since file descriptor 8 is used for
completions when $_ARGCOMPLETE_STDOUT_FILENAME is not set, use the
os.fdopen mock to handle both and simplify the code.

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>